// app/src/generated/models/Csa_notesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Csa_notesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Csa_notesstatecode = keyof typeof Csa_notesstatecode;
export const Csa_notesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Csa_notesstatuscode = keyof typeof Csa_notesstatuscode;

export interface Csa_notesBase {
  csa_name?: string;
  csa_noteid: string;
  "csa_TaskId@odata.bind": string;
  csa_text: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Csa_notesstatecode;
  statuscode?: Csa_notesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Csa_notes extends Csa_notesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  csa_taskidname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  csa_taskid?: object;
  _csa_taskid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}